UIEH-630: Move agreements section to a separate feature component. Apply it for resource show page
