+++ src/components/resource/resource-show.js
@@ -8,7 +8,6 @@ import update from 'lodash/fp/update';
 import set from 'lodash/fp/set';
 
 import {
-  Pluggable,
   IfPermission,
   withStripes,
 } from '@folio/stripes-core';
@@ -31,7 +30,7 @@ import ExternalLink from '../external-link/external-link';
 import IdentifiersList from '../identifiers-list';
 import ContributorsList from '../contributors-list';
 import CoverageDateList from '../coverage-date-list';
-import AgreementsList from '../agreements-list';
+import AgreementsSection from '../../features';
 import {
   isBookPublicationType,
   isValidCoverageList,
@@ -46,8 +45,6 @@ import ProxyDisplay from '../proxy-display';
 
 class ResourceShow extends Component {
   static propTypes = {
-    agreements: PropTypes.object,
-    getAgreements: PropTypes.func.isRequired,
     isFreshlySaved: PropTypes.bool,
     model: PropTypes.object.isRequired,
     onAddAgreement: PropTypes.object.isRequired,
@@ -198,39 +195,6 @@ class ResourceShow extends Component {
     );
   }
 
-  getAgreementsSectionHeader = () => {
-    return (
-      <Headline
-        size="large"
-        tag="h3"
-      >
-        <FormattedMessage id="ui-eholdings.agreements" />
-      </Headline>
-    );
-  }
-
-  renderFindAgreementTrigger = (props) => {
-    return (
-      <Button {...props}>
-        <FormattedMessage id="ui-eholdings.add" />
-      </Button>
-    );
-  }
-
-  getAgreementsSectionButtons() {
-    const {
-      onAddAgreement,
-    } = this.props;
-
-    return (
-      <Pluggable
-        dataKey="resource-show-find-agreement"
-        type="find-agreement"
-        renderTrigger={this.renderFindAgreementTrigger}
-        onAgreementSelected={onAddAgreement}
-      />
-    );
-  }
 
   render() {
     let {
@@ -240,8 +204,6 @@ class ResourceShow extends Component {
       tagsModel,
       updateEntityTags,
       updateFolioTags,
-      agreements,
-      getAgreements,
     } = this.props;
 
     let {
@@ -607,18 +569,12 @@ class ResourceShow extends Component {
                 )}
               </Accordion>
 
-              <Accordion
+              <AgreementsSection
                 id="resourceShowAgreements"
-                label={this.getAgreementsSectionHeader()}
-                open={sections.resourceShowAgreements}
-                displayWhenOpen={this.getAgreementsSectionButtons()}
+                referenceId={model.id}
+                isOpen={sections.resourceShowAgreements}
                 onToggle={this.handleSectionToggle}
-              >
-                <AgreementsList
-                  getAgreements={getAgreements}
-                  agreements={agreements}
-                />
-              </Accordion>
+              />
             </div>
           )}
         />
