+++ src/features/agreements-section/index.js
@@ -1,0 +1,1 @@
+export { default } from './agreements-section';
